+++ src/components/circle/index.tsx
@@ -2,7 +2,11 @@ import * as React from 'react'
 import classNames from 'classnames'
 
 export function Circle (props: any) {
-    const classes = classNames('circle_default', props.className)
+    const { className, type } = props;
+    const prefixCls = 'dtc-circle';
+    const classes = classNames(`${prefixCls}-default`, className, {
+        [`${prefixCls}-${type}`]: type
+    })
     const { ...other } = props;
     return (
         <div