+++ src/stories/index.stories.tsx
@@ -5,6 +5,10 @@ import MarkdownRender from '../components/markdown-render';
 import { notShowProps } from './func';
 import './style';
 const html = require('../../CHANGELOG.md');
+const readmeHtml = require('../../README.md');
+const docHtml = require('./markdown/componentDoc.md');
+const devlopHtml = require('./markdown/componentDev.md');
+const planHtml = require('./markdown/plan.md');
 const { name, repository, version } = require('../../package.json');
 const stories = storiesOf('综述', module)
 console.log('%c欢迎使用 dt-react-component\n使用过程中如有问题欢迎联系 qingyi@dtstack.com ', 'color:#2517b1')
@@ -43,15 +47,65 @@ stories
                 <a rel="noopener noreferrer" target='_blank' href='http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/issues'> Issue</a> 或{' '}
                 <a rel="noopener noreferrer" target='_blank' href='http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/merge_requests'>Pull Request</a>
             </p>
+            <h2>FAQ</h2>
+            <p>使用过程中如有问题欢迎沟通～～</p>
         </article>
     ), {
         ...notShowProps(false, [FaGitlab])
+    })
+    .add(`快速上手`, () => {
+        return (
+            <div className='story_wrapper'>
+                <MarkdownRender
+                    text={`${readmeHtml && readmeHtml.default}`}
+                    dark={false}
+                />
+            </div>
+        )
+    }, {
+        ...notShowProps(false, [MarkdownRender])
+    })
+    .add(`组件文档规范`, () => {
+        return (
+            <div className='story_wrapper'>
+                <MarkdownRender
+                    text={`${docHtml && docHtml.default}`}
+                    dark={false}
+                />
+            </div>
+        )
+    }, {
+        ...notShowProps(false, [MarkdownRender])
+    })
+    .add(`组件开发`, () => {
+        return (
+            <div className='story_wrapper'>
+                <MarkdownRender
+                    text={`${devlopHtml && devlopHtml.default}`}
+                    dark={false}
+                />
+            </div>
+        )
+    }, {
+        ...notShowProps(false, [MarkdownRender])
+    })
+    .add(`Todo/Plan`, () => {
+        return (
+            <div className='story_wrapper'>
+                <MarkdownRender
+                    text={`${planHtml && planHtml.default}`}
+                    dark={false}
+                />
+            </div>
+        )
+    }, {
+        ...notShowProps(false, [MarkdownRender])
     })
     .add(`更新日志`, () => (
         <div className='story_wrapper'>
             <MarkdownRender
                 text={`${html && html.default}`}
-                dark={true}
+                dark={false}
             />
         </div>
     ), {