+++ CHANGELOG.md
@@ -1,12 +1,27 @@
-# 1.0.0 (2020-06-02)
+# 1.0.0 (2020-06-15)
+
+
+### Bug Fixes
+
+* **component:** fix component ([fdd6405](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/fdd6405))
+* **contextmenu:** fix contextmenu ([f78d691](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/f78d691))
+* **docs:** update plan ([1128fea](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/1128fea))
+* **npm:** fix npmignore ([fca883b](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/fca883b))
+* **script:** fix release scripts ([8909cd2](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/8909cd2))
+* **stories:** fix stories ([59d9857](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/59d9857))
 
 
 ### Features
 
-* **component:** add components ([be532fd](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/be532fd))
-* **component:** add table cell component ([16f6772](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/16f6772))
-* **global:** optimze ([25ee271](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/25ee271))
-* **stories:** add index stories ([5f9392d](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/5f9392d))
+* **component:** add context menu ([88ff5b3](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/88ff5b3))
+* **component:** add not found component and update goback ([91587e9](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/91587e9))
+* **component:** add slidepane component ([c0b9be6](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/c0b9be6))
+* **component:** circle component ([505a09e](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/505a09e))
+* **component:** slidePane ([151ad97](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/151ad97))
+* **component:** update context menu and stories sort ([dc26b84](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/dc26b84))
+* **doc:** add dev doc ([b120e18](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/b120e18))
+* **project:** docs ([6e069e3](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/6e069e3))
+* **stories:** addons and docs update ([b656d63](http://gitlab.prod.dtstack.cn/dt-insight-front/infrastructure/dt-react-component/commit/b656d63))